test optional checks and autodownload

+++ clean.sh
@@ -3,7 +3,13 @@
 for D in /sources/*; do
     if [ -d "${D}" ]; then
         #echo "${D}"
-        if [[ "${D}" != *"/sources/Xorg"*  ]] && [[ "${D}" != *"/sources/kf5"*  ]] && [[ "${D}" != *"/sources/plasma"*  ]]; then
+        if [[ "${D}" != *"/sources/Xorg"*  ]]      \
+        && [[ "${D}" != *"/sources/kf5"*  ]]       \
+        && [[ "${D}" != *"/sources/plasma"*  ]]    \
+        && [[ "${D}" != *"/sources/games"*  ]]     \
+        && [[ "${D}" != *"/sources/kdeapps"*  ]]   \
+        && [[ "${D}" != *"*.git"*  ]]; 
+          then
               echo "${D}"  # your processing here
               rm -rf ${D}
         fi
